Working display of deforming models
Modified the way updates are handled by the client. Now updates are placed into a queue instead of instantly handled. This allows the client to apply updates in a more controlled manner without worry that one update might come before another due to latency in loading textures.

Also added a new wrapped method to the Legacy Cutter, that converts the non-split vertices of the simulation mesh to the split vertices of the display mesh.